@@ -1,34 +1,42 @@
 {
-    "name":    "phpdocumentor/reflection-docblock",
+    "name": "phpdocumentor/reflection-docblock",
     "description": "With this component, a library can provide support for annotations via DocBlocks or otherwise retrieve information that is embedded in a DocBlock.",
-    "type":    "library",
+    "type": "library",
     "license": "MIT",
     "authors": [
         {
             "name": "Mike van Riel",
             "email": "me@mikevanriel.com"
+        },
+        {
+            "name": "Jaap van Otterdijk",
+            "email": "account@ijaap.nl"
         }
     ],
     "require": {
-        "php": "^7.0",
-        "phpdocumentor/reflection-common": "^1.0.0",
-        "phpdocumentor/type-resolver": "^0.4.0",
-        "webmozart/assert": "^1.0"
-    },
-    "autoload": {
-        "psr-4": {"phpDocumentor\\Reflection\\": ["src/"]}
-    },
-    "autoload-dev": {
-        "psr-4": {"phpDocumentor\\Reflection\\": ["tests/unit"]}
+        "php": "^7.2",
+        "phpdocumentor/type-resolver": "^1.0",
+        "webmozart/assert": "^1",
+        "phpdocumentor/reflection-common": "^2.0",
+        "ext-filter": "^7.1"
     },
     "require-dev": {
-        "mockery/mockery": "^1.0",
-        "phpunit/phpunit": "^6.4",
-        "doctrine/instantiator": "~1.0.5"
+        "mockery/mockery": "^1",
+        "doctrine/instantiator": "^1"
+    },
+    "autoload": {
+        "psr-4": {
+            "phpDocumentor\\Reflection\\": "src"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "phpDocumentor\\Reflection\\": "tests/unit"
+        }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "4.x-dev"
+            "dev-master": "5.x-dev"
         }
     }
 }
