@@ -1,27 +1,35 @@
 {
-    "name":    "phpdocumentor/type-resolver",
-    "type":    "library",
+    "name": "phpdocumentor/type-resolver",
+    "description": "A PSR-5 based resolver of Class names, Types and Structural Element Names",
+    "type": "library",
     "license": "MIT",
     "authors": [
-        {"name": "Mike van Riel", "email": "me@mikevanriel.com"}
+        {
+          "name": "Mike van Riel",
+          "email": "me@mikevanriel.com"
+        }
     ],
     "require": {
-        "php": "^5.5 || ^7.0",
-        "phpdocumentor/reflection-common": "^1.0"
-    },
-    "autoload": {
-        "psr-4": {"phpDocumentor\\Reflection\\": ["src/"]}
-    },
-    "autoload-dev": {
-        "psr-4": {"phpDocumentor\\Reflection\\": ["tests/unit"]}
+        "php": "^7.2",
+        "phpdocumentor/reflection-common": "^2.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.2||^4.8.24",
-        "mockery/mockery": "^0.9.4"
+        "mockery/mockery": "~1",
+        "ext-tokenizer": "^7.2"
+    },
+    "autoload": {
+        "psr-4": {
+            "phpDocumentor\\Reflection\\": "src"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "phpDocumentor\\Reflection\\": "tests/unit"
+        }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0.x-dev"
+            "dev-master": "1.x-dev"
         }
     }
 }
