@@ -16,26 +16,33 @@
         }
     ],
     "require": {
-        "php": "^5.5.9|>=7.0.8",
+        "php": "^7.1.3",
         "symfony/polyfill-mbstring": "~1.0",
-        "symfony/debug": "~2.8|~3.0|~4.0"
+        "symfony/polyfill-php73": "^1.8",
+        "symfony/service-contracts": "^1.1|^2"
     },
     "require-dev": {
-        "symfony/config": "~3.3|~4.0",
-        "symfony/event-dispatcher": "~2.8|~3.0|~4.0",
-        "symfony/dependency-injection": "~3.4|~4.0",
-        "symfony/lock": "~3.4|~4.0",
-        "symfony/process": "~3.3|~4.0",
+        "symfony/config": "^3.4|^4.0|^5.0",
+        "symfony/event-dispatcher": "^4.3",
+        "symfony/dependency-injection": "^3.4|^4.0|^5.0",
+        "symfony/lock": "^4.4|^5.0",
+        "symfony/process": "^3.4|^4.0|^5.0",
+        "symfony/var-dumper": "^4.3|^5.0",
         "psr/log": "~1.0"
+    },
+    "provide": {
+        "psr/log-implementation": "1.0"
     },
     "suggest": {
         "symfony/event-dispatcher": "",
         "symfony/lock": "",
         "symfony/process": "",
-        "psr/log-implementation": "For using the console logger"
+        "psr/log": "For using the console logger"
     },
     "conflict": {
         "symfony/dependency-injection": "<3.4",
+        "symfony/event-dispatcher": "<4.3|>=5",
+        "symfony/lock": "<4.4",
         "symfony/process": "<3.3"
     },
     "autoload": {
@@ -47,7 +54,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "3.4-dev"
+            "dev-master": "4.4-dev"
         }
     }
 }
