@@ -16,12 +16,16 @@
         }
     ],
     "require": {
-        "php": "^5.5.9|>=7.0.8",
+        "php": "^7.1.3",
         "symfony/polyfill-ctype": "~1.8",
         "symfony/polyfill-mbstring": "~1.0"
     },
     "require-dev": {
-        "symfony/css-selector": "~2.8|~3.0|~4.0"
+        "symfony/css-selector": "^3.4|^4.0|^5.0",
+        "masterminds/html5": "^2.6"
+    },
+    "conflict": {
+        "masterminds/html5": "<2.6"
     },
     "suggest": {
         "symfony/css-selector": ""
@@ -35,7 +39,7 @@
     "minimum-stability": "dev",
     "extra": {
         "branch-alias": {
-            "dev-master": "3.4-dev"
+            "dev-master": "4.4-dev"
         }
     }
 }
