@@ -16,35 +16,43 @@ return [
         'basePath'            => dirname(__DIR__),
         'controllerNamespace' => 'backend\controllers',
         'bootstrap'           => ['log'],
-        'modules'             => [],
+        'modules'             => [
+                'grapesjs' => [
+                        'class'             => \thecodeholic\yii2grapesjs\Module::class,
+                        // custom placeholder variables which will be added into richtext
+                        // default is empty array
+                        'grapesJsVariables' => []
+                ],
+        ],
         'components'          => [
                 'request'    => [
-                        'csrfParam' => '_csrf-backend',
-                        'class'     => 'common\components\Request',
-                        'web'       => '/backend/web',
-                        'adminUrl'  => $urlStr,
+                        'enableCsrfValidation' => true,
+                        'csrfParam'            => '_csrf-backend',
+                        'class'                => 'common\components\Request',
+                        'web'                  => '/backend/web',
+                        'adminUrl'             => $urlStr,
                         //  'baseUrl'   => $baseUrl,
                 ],
                 'urlManager' => [
                     //     'baseUrl' => $baseUrl,
                     'rules' => [
-                            'content/page/<page:[a-zA-Z0-9-]+>/' => 'content/page',
+                            'content/page/<page:[a-zA-Z0-9-]+>/'             => 'content/page',
                             'content/fetch-categories/<type:[a-zA-Z0-9-]+>/' => 'content/fetch-categories',
-                            'settings/edit/<id:[a-zA-Z0-9-]+>/'         => 'settings/',
-                            'clients/edit/<id:[a-zA-Z0-9-]+>/'          => 'clients/',
-                            'users/edit/<id:[a-zA-Z0-9-]+>/'            => 'users/edit/',
-                            'blog/view-comment/<id:[a-zA-Z0-9-]+>/'     => 'blog/view-comment/',
-                            'testimonials/edit/<id:[a-zA-Z0-9-]+>/'     => 'testimonials/',
-                            'team/edit/<id:[a-zA-Z0-9-]+>/'             => 'team/',
-                            'package/category/<id:[a-zA-Z0-9-]+>/'      => 'package/category/',
-                            'amenities/edit/<id:[a-zA-Z0-9-]+>/'        => 'amenities/',
-                            'slider/post/<id:[a-zA-Z0-9-]+>/'           => 'slider/post',
-                            'faq/edit/<id:[a-zA-Z0-9-]+>/'              => 'faq/',
-                            'package/category-edit/<id:[a-zA-Z0-9-]+>/' => 'package/category',
-                            'package/edit-city/<id:[a-zA-Z0-9-]+>/'     => 'package/cities',
-                            'package/city-post/<id:[a-zA-Z0-9-]+>/'     => 'package/city-post',
-                            'package/post/<id:[a-zA-Z0-9-]+>/'          => 'package/post',
-                            'services/edit/<id:[a-zA-Z0-9-]+>/'         => 'services/',
+                            'settings/edit/<id:[a-zA-Z0-9-]+>/'              => 'settings/',
+                            'clients/edit/<id:[a-zA-Z0-9-]+>/'               => 'clients/',
+                            'users/edit/<id:[a-zA-Z0-9-]+>/'                 => 'users/edit/',
+                            'blog/view-comment/<id:[a-zA-Z0-9-]+>/'          => 'blog/view-comment/',
+                            'testimonials/edit/<id:[a-zA-Z0-9-]+>/'          => 'testimonials/',
+                            'team/edit/<id:[a-zA-Z0-9-]+>/'                  => 'team/',
+                            'package/category/<id:[a-zA-Z0-9-]+>/'           => 'package/category/',
+                            'amenities/edit/<id:[a-zA-Z0-9-]+>/'             => 'amenities/',
+                            'slider/post/<id:[a-zA-Z0-9-]+>/'                => 'slider/post',
+                            'faq/edit/<id:[a-zA-Z0-9-]+>/'                   => 'faq/',
+                            'package/category-edit/<id:[a-zA-Z0-9-]+>/'      => 'package/category',
+                            'package/edit-city/<id:[a-zA-Z0-9-]+>/'          => 'package/cities',
+                            'package/city-post/<id:[a-zA-Z0-9-]+>/'          => 'package/city-post',
+                            'package/post/<id:[a-zA-Z0-9-]+>/'               => 'package/post',
+                            'services/edit/<id:[a-zA-Z0-9-]+>/'              => 'services/',
                             [
                                     'pattern'      => 'sections/pages/<page:[a-zA-Z0-9-]+>',
                                     'route'        => 'sections/pages/',
