@@ -84,8 +84,12 @@ $this->registerCssFile(Yii::$app->request->baseUrl . '/resources/libs/datatables
                                  </td>
                                  <td> <?php echo(isset($post['title']) && !empty($post['title']) ? trim($post['title']) : ''); ?></td>
                                  <td> <?php echo(isset($post['category']) && !empty($post['category']) ? trim($post['category']) : ''); ?></td>
-                                 <td> <?php echo(isset($post['info']['visibility']) && ($post['info']['visibility'] == 1) ? 'Visible' : 'Hidden'); ?></td>
-                                 <td> <?php echo(isset($post['author']) && !empty($post['author']) ? ucwords(trim($post['author'])) : ''); ?></td>
+                                 <td>
+                                    <input type="checkbox" id="switch3<?= $post['id']?>" class="status" switch="bool" data-id = "<?php echo \common\components\Misc::encodeUrl($post['blog_id']); ?>" data-tab = "Blog"
+                                            <?php echo(isset($post['info']['is_active']) && ($post['info']['is_active'] == 1) ? 'checked' : ''); ?>>
+                                    <label for="switch3<?= $post['id']?>" data-on-label="Visible"
+                                           data-off-label="Hidden"></label>
+                                 </td>                                 <td> <?php echo(isset($post['author']) && !empty($post['author']) ? ucwords(trim($post['author'])) : ''); ?></td>
 
                                  <td>
                                      <?php
