@@ -33,11 +33,7 @@ class HelperBlog extends Component {
         else {
             $model = new Blog();
         }
-echo '<pre>';
-print_r($model);
-echo '</pre>';
-die;
-        $model->visibility = $data['visibility'];
+        $model->is_active = $data['visibility'];
         if (isset($image['name']) && $image['name'] != '') {
             if ($model->image != '') {
                 Misc::delete_file($model->image, 'image');
