@@ -5,15 +5,12 @@ namespace common\models\generated;
 use Yii;
 
 /**
- * This is the model class for table "package_category_translation".
+ * This is the model class for table "{{%package_category_translation}}".
  *
  * @property int $id
  * @property int $package_category_id
- * @property string $lamguage_code
+ * @property string $language_code
  * @property string $name
- *
- * @property PackageCategory $packageCategory
- * @property Language $lamguageCode
  */
 class PackageCategoryTranslation extends \yii\db\ActiveRecord
 {
@@ -22,7 +19,7 @@ class PackageCategoryTranslation extends \yii\db\ActiveRecord
      */
     public static function tableName()
     {
-        return 'package_category_translation';
+        return '{{%package_category_translation}}';
     }
 
     /**
@@ -31,11 +28,9 @@ class PackageCategoryTranslation extends \yii\db\ActiveRecord
     public function rules()
     {
         return [
-            [['package_category_id', 'lamguage_code', 'name'], 'required'],
+            [['package_category_id', 'language_code', 'name'], 'required'],
             [['package_category_id'], 'integer'],
-            [['lamguage_code', 'name'], 'string', 'max' => 200],
-            [['package_category_id'], 'exist', 'skipOnError' => true, 'targetClass' => PackageCategory::className(), 'targetAttribute' => ['package_category_id' => 'id']],
-            [['lamguage_code'], 'exist', 'skipOnError' => true, 'targetClass' => Language::className(), 'targetAttribute' => ['lamguage_code' => 'code']],
+            [['language_code', 'name'], 'string', 'max' => 200],
         ];
     }
 
@@ -47,24 +42,8 @@ class PackageCategoryTranslation extends \yii\db\ActiveRecord
         return [
             'id' => 'ID',
             'package_category_id' => 'Package Category ID',
-            'lamguage_code' => 'Lamguage Code',
+            'language_code' => 'Language Code',
             'name' => 'Name',
         ];
-    }
-
-    /**
-     * @return \yii\db\ActiveQuery
-     */
-    public function getPackageCategory()
-    {
-        return $this->hasOne(PackageCategory::className(), ['id' => 'package_category_id']);
-    }
-
-    /**
-     * @return \yii\db\ActiveQuery
-     */
-    public function getLamguageCode()
-    {
-        return $this->hasOne(Language::className(), ['code' => 'lamguage_code']);
     }
 }
